requirements.txt
# Conda packages:
# python=3.9.9
# tensorflow-gpu>=2.7.0
# tensorflow-addons>=0.13.0

# PIP packages:
tensorflow>=2.8.0
tensorflow-addons>=0.16.1
jupyter>=1.0.0
pandas>=1.4.0
matplotlib>=3.5.0
scipy>=1.7.3
pyyaml>=6.0
Pillow>=8.4.0
opencv-python>=4.5.4
pydantic>=1.8.2
pydantic-cli>=4.2.1
pydantic-yaml>=0.5.1
plyfile>=0.7.4
keras-applications>=1.0.8
Cython>=0.29.26
pyopengl>=3.1.5
pyopengl-accelerate>=3.1.5
glfw>=2.5.0
scikit-learn>=1.0.2
imgaug>=0.4.0
tqdm>=4.64.0

# Built from source
# pymesh2>=0.3

